Clean compiles on report and print

+++ settings.h
@@ -66,7 +66,7 @@ typedef struct {
   float acceleration;
   float junction_deviation;
   uint8_t flags;  // Contains default boolean settings
-  uint8_t homing_dir_mask;
+  uint32_t homing_dir_mask;
   float homing_feed_rate;
   float homing_seek_rate;
   uint16_t homing_debounce_delay;
